Refactor test utilities and consolidate common assertions
- Moved common assertion functions from individual test files into dedicated utility files under `testUtil`.
- Created utility functions for color assertions, completion assertions, formatting assertions, linting assertions, navigation assertions, node assertions, and parser assertions.
- Updated all relevant test files to import and use the new utility functions, improving code organization and reducing duplication.
- Removed redundant code and streamlined the test structure for better maintainability.
