+++ package.json
@@ -39,8 +39,8 @@
     "remove-sourcemap-refs": "node ./build/remove-sourcemap-refs.js",
     "watch": "tsc -w -p ./src",
     "test": "npm run compile && npm run node-test",
-    "node-test": "node --enable-source-maps --test --test-isolation=none 'lib/esm/test/**/*.test.js'",
-    "coverage": "npm run compile && npx nyc --reporter=html --reporter=text node --test --test-isolation=none 'lib/esm/test/**/*.test.js'",
+    "node-test": "node --enable-source-maps --test 'lib/esm/test/**/*.test.js'",
+    "coverage": "npm run compile && npx nyc --reporter=html --reporter=text node --test 'lib/esm/test/**/*.test.js'",
     "lint": "node --test ./build/lint.test.ts",
     "update-data": "npm install @vscode/web-custom-data -D && node ./build/generateData.js",
     "install-types-next": "npm install vscode-languageserver-types@next -f -S && npm install vscode-languageserver-textdocument@next -f -S",
