+++ src/test/css/completion.test.ts
@@ -18,161 +18,9 @@ import { getDocumentContext } from '../testUtil/documentContext.js';
 import { URI } from 'vscode-uri';
 import { getFsProvider } from '../testUtil/fsProvider.js';
 import { TextEdit } from 'vscode-languageserver-types';
+import { ExpectedCompetions, testCompletionFor } from '../testUtil/completion.js';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
-
-export interface ItemDescription {
-	label: string;
-	detail?: string;
-	documentation?: string | MarkupContent | null;
-	/**
-	 * Only test that the documentation includes the substring
-	 */
-	documentationIncludes?: string;
-	kind?: CompletionItemKind;
-	insertTextFormat?: InsertTextFormat;
-	resultText?: string;
-	notAvailable?: boolean;
-	command?: Command;
-	sortText?: string;
-}
-
-export function assertCompletion(completions: CompletionList, expected: ItemDescription, document: TextDocument) {
-	const matches = completions.items.filter(completion => {
-		return completion.label === expected.label;
-	});
-	if (expected.notAvailable) {
-		assert.equal(matches.length, 0, expected.label + " should not be present");
-	} else {
-		assert.equal(matches.length, 1, expected.label + " should only existing once: Actual: " + completions.items.map(c => c.label).join(', '));
-	}
-
-	const match = matches[0];
-	if (expected.detail) {
-		assert.equal(match.detail, expected.detail);
-	}
-	if (expected.documentation) {
-		assert.deepEqual(match.documentation, expected.documentation);
-	}
-	if (expected.documentationIncludes) {
-		assert.ok(match.documentation !== undefined);
-		if (typeof match.documentation === 'string') {
-			assert.ok(match.documentation.indexOf(expected.documentationIncludes) !== -1);
-		} else {
-			assert.ok(match.documentation!.value.indexOf(expected.documentationIncludes) !== -1);
-		}
-	}
-	if (expected.kind) {
-		assert.equal(match.kind, expected.kind);
-	}
-	if (expected.resultText && match.textEdit) {
-		const edit = TextEdit.is(match.textEdit) ? match.textEdit : TextEdit.replace(match.textEdit.replace, match.textEdit.newText);
-		assert.equal(TextDocument.applyEdits(document, [edit]), expected.resultText);
-	}
-	if (expected.insertTextFormat) {
-		assert.equal(match.insertTextFormat, expected.insertTextFormat);
-	}
-	if (expected.command) {
-		assert.deepEqual(match.command, expected.command);
-	}
-	if (expected.sortText) {
-		assert.equal(match.sortText, expected.sortText);
-	}
-};
-
-export type ExpectedCompetions = {
-	count?: number;
-	items?: ItemDescription[];
-	participant?: {
-		onProperty?: PropertyCompletionContext[];
-		onPropertyValue?: PropertyValueCompletionContext[];
-		onURILiteralValue?: URILiteralCompletionContext[];
-		onImportPath?: ImportPathCompletionContext[];
-		onMixinReference?: MixinReferenceCompletionContext[];
-	};
-};
-
-export async function testCompletionFor(
-	value: string,
-	expected: ExpectedCompetions,
-	settings: LanguageSettings = {
-		completion: {
-			triggerPropertyValueCompletion: true,
-			completePropertyWithSemicolon: false
-		}
-	},
-	testUri: string = 'test://test/test.css',
-	workspaceFolderUri: string = 'test://test',
-	customData: ICSSDataProvider[] = [],
-) {
-	const offset = value.indexOf('|');
-	assert.ok(offset !== -1, '| missing in ' + value);
-	value = value.substr(0, offset) + value.substr(offset + 1);
-
-	const actualPropertyContexts: PropertyCompletionContext[] = [];
-	const actualPropertyValueContexts: PropertyValueCompletionContext[] = [];
-	const actualURILiteralValueContexts: URILiteralCompletionContext[] = [];
-	const actualImportPathContexts: ImportPathCompletionContext[] = [];
-	const actualMixinReferenceContexts: MixinReferenceCompletionContext[] = [];
-
-	const lang = path.extname(testUri).substr(1);
-	const lsOptions = { fileSystemProvider: getFsProvider() };
-	let ls;
-	if (lang === 'scss') {
-		ls = getSCSSLanguageService(lsOptions);
-	} else if (lang === 'less') {
-		ls = getLESSLanguageService(lsOptions);
-	} else {
-		ls = getCSSLanguageService(lsOptions);
-	}
-	ls.setDataProviders(true, customData);
-	ls.configure(settings);
-
-	if (expected.participant) {
-		ls.setCompletionParticipants([
-			{
-				onCssProperty: context => actualPropertyContexts.push(context),
-				onCssPropertyValue: context => actualPropertyValueContexts.push(context),
-				onCssURILiteralValue: context => actualURILiteralValueContexts.push(context),
-				onCssImportPath: context => actualImportPathContexts.push(context),
-				onCssMixinReference: context => actualMixinReferenceContexts.push(context)
-			}
-		]);
-	}
-
-	const document = TextDocument.create(testUri, lang, 0, value);
-	const position = document.positionAt(offset);
-	const jsonDoc = ls.parseStylesheet(document);
-
-	const context = getDocumentContext(workspaceFolderUri);
-
-	const list = await ls.doComplete2(document, position, jsonDoc, context);
-	if (typeof expected.count === 'number') {
-		assert.equal(list.items.length, expected.count);
-	}
-	if (expected.items) {
-		for (const item of expected.items) {
-			assertCompletion(list, item, document);
-		}
-	}
-	if (expected.participant) {
-		if (expected.participant.onProperty) {
-			assert.deepEqual(actualPropertyContexts, expected.participant.onProperty);
-		}
-		if (expected.participant.onPropertyValue) {
-			assert.deepEqual(actualPropertyValueContexts, expected.participant.onPropertyValue);
-		}
-		if (expected.participant.onURILiteralValue) {
-			assert.deepEqual(actualURILiteralValueContexts, expected.participant.onURILiteralValue);
-		}
-		if (expected.participant.onImportPath) {
-			assert.deepEqual(actualImportPathContexts, expected.participant.onImportPath);
-		}
-		if (expected.participant.onMixinReference) {
-			assert.deepEqual(actualMixinReferenceContexts, expected.participant.onMixinReference);
-		}
-	}
-};
 
 suite('CSS - Completion', () => {
