+++ src/test/css/lint.test.ts
@@ -14,24 +14,10 @@ import { TextDocument } from '../../cssLanguageTypes.js';
 import { SCSSParser } from '../../parser/scssParser.js';
 import { LESSParser } from '../../parser/lessParser.js';
 import { CSSDataManager } from '../../languageFacts/dataManager.js';
+import { assertEntries } from '../testUtil/lint.js';
 
 const cssDataManager = new CSSDataManager({ useDefaultDataProvider: true });
 
-export function assertEntries(node: Node, document: TextDocument, expectedRules: IRule[], expectedMessages: string[] | undefined = undefined, settings = new LintConfigurationSettings()): void {
-
-	const entries = LintVisitor.entries(node, document, settings, cssDataManager, Level.Error | Level.Warning | Level.Ignore);
-	const message = `Did not find all linting error. expected: [${expectedRules.map(e => e.id).join(', ')}], actual: [${entries.map(e => e.getMessage()).join(', ')}]`;
-
-	assert.equal(entries.length, expectedRules.length, message);
-
-	for (const entry of entries) {
-		const index = expectedRules.indexOf(entry.getRule());
-		assert.ok(index !== -1, `${entry.getRule().id} found but not expected (${expectedRules.map(r => r.id).join(', ')})`);
-		if (expectedMessages) {
-			assert.equal(entry.getMessage(), expectedMessages[index]);
-		}
-	}
-}
 const parsers = [new Parser(), new LESSParser(), new SCSSParser()];
 
 function assertStyleSheet(input: string, ...rules: Rule[]): void {
