+++ src/test/css/navigation.test.ts
@@ -20,169 +20,12 @@ import {
 import { URI } from 'vscode-uri';
 import { getFsProvider } from '../testUtil/fsProvider.js';
 import { getDocumentContext } from '../testUtil/documentContext.js';
+import {
+	assertColorPresentations, assertColorSymbols, assertDocumentSymbols, assertHighlights, assertLinks,
+	assertScopeBuilding, assertScopesAndSymbols, assertSymbolInfos, assertSymbolsInScope, getTestResource, newRange
+} from '../testUtil/navigation.js';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
-
-export function assertScopesAndSymbols(ls: LanguageService, input: string, expected: string): void {
-	const global = createScope(ls, input);
-	assert.equal(scopeToString(global), expected);
-}
-
-export function assertHighlights(ls: LanguageService, input: string, marker: string, expectedMatches: number, expectedWrites: number, elementName?: string) {
-	const document = TextDocument.create('test://test/test.css', 'css', 0, input);
-
-	const stylesheet = ls.parseStylesheet(document);
-	assertNoErrors(stylesheet);
-
-	const index = input.indexOf(marker) + marker.length;
-	const position = document.positionAt(index);
-
-	const highlights = ls.findDocumentHighlights(document, position, stylesheet);
-	assert.equal(highlights.length, expectedMatches, input);
-
-	let nWrites = 0;
-	for (const highlight of highlights) {
-		if (highlight.kind === DocumentHighlightKind.Write) {
-			nWrites++;
-		}
-		const range = highlight.range;
-		const start = document.offsetAt(range.start), end = document.offsetAt(range.end);
-		assert.equal(document.getText().substring(start, end), elementName || marker);
-	}
-	assert.equal(nWrites, expectedWrites, input);
-}
-
-
-
-export async function assertLinks(ls: LanguageService, input: string, expected: DocumentLink[], lang: string = 'css', testUri?: string, workspaceFolder?: string) {
-	const document = TextDocument.create(testUri || `test://test/test.${lang}`, lang, 0, input);
-
-	const stylesheet = ls.parseStylesheet(document);
-
-	const links = await ls.findDocumentLinks2(document, stylesheet, getDocumentContext(workspaceFolder || 'test://test'));
-	assert.deepEqual(links, expected);
-}
-
-export function assertSymbolInfos(ls: LanguageService, input: string, expected: SymbolInformation[], lang: string = 'css') {
-	const document = TextDocument.create(`test://test/test.${lang}`, lang, 0, input);
-
-	const stylesheet = ls.parseStylesheet(document);
-
-	const symbols = ls.findDocumentSymbols(document, stylesheet);
-	assert.deepEqual(symbols, expected);
-}
-
-export function assertDocumentSymbols(ls: LanguageService, input: string, expected: DocumentSymbol[], lang: string = 'css') {
-	const document = TextDocument.create(`test://test/test.${lang}`, lang, 0, input);
-
-	const stylesheet = ls.parseStylesheet(document);
-
-	const symbols = ls.findDocumentSymbols2(document, stylesheet);
-	assert.deepEqual(symbols, expected);
-}
-
-export function assertColorSymbols(ls: LanguageService, input: string, ...expected: ColorInformation[]) {
-	const document = TextDocument.create('test://test/test.css', 'css', 0, input);
-
-	const stylesheet = ls.parseStylesheet(document);
-	const result = ls.findDocumentColors(document, stylesheet);
-	assert.deepEqual(result, expected);
-}
-
-export function assertColorPresentations(ls: LanguageService, color: Color, ...expected: string[]) {
-	const document = TextDocument.create('test://test/test.css', 'css', 0, '');
-
-	const stylesheet = ls.parseStylesheet(document);
-	const range = newRange(1, 2);
-	const result = ls.getColorPresentations(document, stylesheet, color, range);
-	assert.deepEqual(result.map(r => r.label), expected);
-	assert.deepEqual(result.map(r => r.textEdit), expected.map(l => TextEdit.replace(range, l)));
-}
-
-export function assertSymbolsInScope(ls: LanguageService, input: string, offset: number, ...selections: { name: string; type: nodes.ReferenceType }[]): void {
-
-	const global = createScope(ls, input);
-
-	const scope = global.findScope(offset)!;
-
-	const getErrorMessage = function (name: string) {
-		let all = 'symbol ' + name + ' not found. In scope: ';
-		scope.getSymbols().forEach((sym) => { all += (sym.name + ' '); });
-		return all;
-	};
-
-	for (let i = 0; i < selections.length; i++) {
-		const selection = selections[i];
-		const sym = scope.getSymbol(selection.name, selection.type) || global.getSymbol(selection.name, selection.type);
-		assert.ok(!!sym, getErrorMessage(selection.name));
-	}
-}
-
-export function assertScopeBuilding(ls: LanguageService, input: string, ...scopes: { offset: number; length: number; }[]): void {
-
-	const global = createScope(ls, input);
-
-	function assertChildren(scope: Scope): void {
-
-		scope.children.forEach((scope) => {
-
-			// check bounds
-			const expected = scopes.shift()!;
-			assert.equal(scope.offset, expected.offset);
-			assert.equal(scope.length, expected.length);
-
-			// recursive descent
-			assertChildren(scope);
-		});
-	}
-
-	assertChildren(global);
-
-	assert.equal(scopes.length, 0, 'remaining scopes: ' + scopes.join());
-}
-
-export function getTestResource(path: string) {
-	return URI.file(join(__dirname, '../../../../test/linksTestFixtures', path)).toString(true);
-}
-
-function scopeToString(scope: Scope): string {
-	let str = '';
-	const symbols = scope.getSymbols();
-	for (let index = 0; index < symbols.length; index++) {
-		if (str.length > 0) {
-			str += ',';
-		}
-		str += symbols[index].name;
-	}
-	const scopes = scope.children;
-	for (let index = 0; index < scopes.length; index++) {
-		if (str.length > 0) {
-			str += ',';
-		}
-		str += ('[' + scopeToString(scopes[index]) + ']');
-	}
-	return str;
-}
-
-function assertNoErrors(stylesheet: Stylesheet): void {
-	const markers = nodes.ParseErrorCollector.entries(<nodes.Stylesheet>stylesheet);
-	if (markers.length > 0) {
-		assert.ok(false, 'node has errors: ' + markers[0].getMessage() + ', offset: ' + markers[0].getNode().offset);
-	}
-}
-
-function createScope(ls: LanguageService, input: string): Scope {
-	const document = TextDocument.create('test://test/test.css', 'css', 0, input);
-
-	const styleSheet = ls.parseStylesheet(document),
-		global = new GlobalScope(),
-		builder = new ScopeBuilder(global);
-
-	assertNoErrors(styleSheet);
-
-	(<nodes.Stylesheet>styleSheet).acceptVisitor(builder);
-	return global;
-}
 
 function getCSSLS() {
 	return getCSSLanguageService({ fileSystemProvider: getFsProvider() });
@@ -526,7 +369,3 @@ suite('CSS - Navigation', () => {
 		});
 	});
 });
-
-export function newRange(start: number, end: number) {
-	return Range.create(Position.create(0, start), Position.create(0, end));
-}
