+++ src/test/scss/formatter.test.ts
@@ -5,7 +5,7 @@
 
 import { suite, test } from 'node:test';
 import { getSCSSLanguageService } from '../../cssLanguageService.js';
-import { assertFormat } from '../css/formatter.test.js';
+import { assertFormat } from '../testUtil/formatter.js';
 
 suite('SCSS - Formatter', () => {
